Move code to test directory
The code is not used in the SDK itself, so sholudn't be part of it.
